Fixed wrong cast in GraqlIT, merged all tests into AnalyticsIT, and renamed to ComputeQueryIT
